# specs/api/openapi.yaml
openapi: 3.1.0
info:
  title: TableFlow API
  version: 0.1.0
  description: >-
    Machine-readable wire contract for implemented TableFlow routes only.
    Policies and planned routes are described in http.md. Routes are added
    here before their implementation lands.
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: http://localhost:8080
    description: Local Go API (also reachable through each frontend's same-origin /api/v1 proxy)
security: []
...
paths:
...
  /api/v1/sessions/anonymous:
    post:
      operationId: startAnonymousSession
      tags: [guest]
      summary: Create or reuse an anonymous bootstrap session
      description: >-
        Requires an allowed PWA Origin; throttled per client IP (60 per 10
        minutes). Returns 200 without Set-Cookie when a valid anonymous
        cookie is already present.
      responses:
        '200':
          description: Existing session reused
          headers:
            Cache-Control: {$ref: '#/components/headers/PrivateNoStore'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AnonymousSession'}
        '201':
          description: New session; sets `__Host-tf_anon` (HttpOnly, Secure, SameSite=Lax, Path=/)
          headers:
            Cache-Control: {$ref: '#/components/headers/PrivateNoStore'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AnonymousSession'}
        '403': {$ref: '#/components/responses/Error'}
        '429': {$ref: '#/components/responses/RateLimited'}
  /api/v1/sessions/capability:
    post:
      operationId: exchangeCapability
      tags: [guest]
      summary: Exchange a queue or dining QR token for a guest session
      description: >-
        The token comes from the URL fragment and is sent only in this body.
        It is not consumed; every diner can exchange the active token. Any
        unusable token (unknown, rotated, revoked, expired, wrong kind)
        returns 422 TOKEN_INVALID. Throttled per client IP (120 per 10 minutes).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [token, kind]
              properties:
                token: {type: string, maxLength: 64}
                kind: {type: string, enum: [queue, visit]}
      responses:
        '201':
          description: Guest session created; sets `__Host-tf_guest` (HttpOnly, Secure, SameSite=Lax, Path=/)
          headers:
            Cache-Control: {$ref: '#/components/headers/PrivateNoStore'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/GuestSession'}
        '400': {$ref: '#/components/responses/Error'}
        '403': {$ref: '#/components/responses/Error'}
        '415': {$ref: '#/components/responses/Error'}
        '422': {$ref: '#/components/responses/Error'}
        '429': {$ref: '#/components/responses/RateLimited'}
  /api/v1/sessions/guest:
    get:
      operationId: getGuestSession
      tags: [guest]
      summary: Current guest session scope
      security: [{guestSession: []}]
      responses:
        '200':
          description: Active guest session
          headers:
            Cache-Control: {$ref: '#/components/headers/PrivateNoStore'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/GuestSession'}
        '401': {$ref: '#/components/responses/Error'}
components:
  securitySchemes:
    guestSession:
      type: apiKey
      in: cookie
      name: __Host-tf_guest
      description: Capability-derived guest session on the PWA origin; valid only for the capability's current generation.
    staffSession:
      type: apiKey
      in: cookie
      name: __Host-tf_staff
      description: Host-only staff session cookie. Mutations also require an allowed admin Origin.
  responses:
    Error:
      description: Error in the common envelope
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
    RateLimited:
      description: Too many attempts
      headers:
        Retry-After:
          description: Seconds until the window resets
          schema: {type: integer, minimum: 1}
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
  headers:
...
    PrivateNoStore:
      description: Authenticated responses are private and never stored.
      schema:
        type: string
        const: private, no-store
  schemas:
...
    Error:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message, request_id, fields]
          properties:
            code:
              type: string
              pattern: '^[A-Z][A-Z0-9_]*$'
              examples: [DEPENDENCY_UNAVAILABLE]
            message:
              type: string
            request_id:
              type: string
            fields:
              type: object
              additionalProperties:
                type: string
...
    GuestSession:
      type: object
      additionalProperties: false
      required: [branch_id, kind, resource_id, expires_at]
      properties:
        branch_id: {type: string, format: uuid}
        kind: {type: string, enum: [queue, visit]}
        resource_id: {type: string, format: uuid}
        expires_at: {type: string, format: date-time}
    AnonymousSession:
      type: object
      additionalProperties: false
      required: [expires_at]
      properties:
        expires_at: {type: string, format: date-time}