# src/pwa/lib/api/types.ts
// Wire types for guest routes in specs/api/openapi.yaml.
export type CapabilityKind = "queue" | "visit";

export type GuestSession = {
  branch_id: string;
  kind: CapabilityKind;
  resource_id: string;
  expires_at: string;
};